+++ Chat Core/MVIRCChatUser.m
@@ -35,7 +35,7 @@
 
 #pragma mark -
 
-- (unsigned long) supportedModes {
+- (NSUInteger) supportedModes {
 	return MVChatUserInvisibleMode;
 }
